attempt to fix test - #1
