+++ test/api/test_cp.py
@@ -231,6 +231,8 @@ class TestComputingProviders:
 
 
 class TestComputingProvidersListByRegion:
+    def setup(self):
+        self.provider = ComputerProvider()
     def test_returns_list_of_computing_providers(self):
         # Mock the requests.post method to return a mock response
         mock_response = Mock()
@@ -241,7 +243,7 @@ class TestComputingProvidersListByRegion:
         requests.post = MagicMock(return_value=mock_response)
 
         # Call the function you're testing
-        result = get_computing_providers_list("region")
+        result = self.provider.get_computing_providers_list("region")
 
         # Assert that the result is a list of dictionaries
         assert isinstance(result, list)
@@ -258,13 +260,13 @@ class TestComputingProvidersListByRegion:
 
             # Call the function under test
             with pytest.raises(SwanHTTPError):
-                get_computing_providers_list("region")
+                self.provider.get_computing_providers_list("region")
 
     def test_raises_value_error_if_region_not_provided(self):
         with pytest.raises(ValueError):
-            get_computing_providers_list("")
+            self.provider.get_computing_providers_list("")
         with pytest.raises(ValueError):
-            get_computing_providers_list(None)
+            self.provider.get_computing_providers_list(None)
 
     def test_raises_swan_http_error(self):
         # Mock the requests.post method to raise an HTTPError
@@ -275,7 +277,7 @@ class TestComputingProvidersListByRegion:
 
             # Call the function under test
             with pytest.raises(SwanHTTPError):
-                get_computing_providers_list("region")
+                self.provider.get_computing_providers_list("region")
 
     def test_raises_exception_for_other_types_of_exceptions(self):
         # Mock the requests.post method to raise an exception
@@ -284,7 +286,7 @@ class TestComputingProvidersListByRegion:
 
         with patch("requests.post", side_effect=mock_post):
             with pytest.raises(Exception):
-                get_computing_providers_list("region")
+                self.provider.get_computing_providers_list("region")
 
     def test_retrieve_collateral_balance_valid_address(self):
         # Mock the requests.get method to return a mock response
@@ -299,7 +301,7 @@ class TestComputingProvidersListByRegion:
             mock_get.return_value = mock_response
 
             # Call the function with a valid computing provider address
-            result = get_collateral_balance("0x1234abcd")
+            result = self.provider.get_collateral_balance("0x1234abcd")
 
             # Assert that the response is as expected
             assert result == {
@@ -322,7 +324,7 @@ class TestComputingProvidersListByRegion:
         with patch("requests.get", side_effect=requests.exceptions.RequestException):
             # Call the function under test
             with pytest.raises(SwanRequestError) as e:
-                get_collateral_balance("0x1234abcd")
+                self.provider.get_collateral_balance("0x1234abcd")
             # Assert that the exception message is correct
             assert (
                 str(e.value)
@@ -342,7 +344,7 @@ class TestComputingProvidersListByRegion:
             mock_get.return_value = mock_response
 
             # Call the function you're testing
-            result = get_collateral_balance("0x1234abcd")
+            result = self.provider.get_collateral_balance("0x1234abcd")
 
             # Assert that requests.get was called with the correct endpoint
             mock_get.assert_called_once_with(
@@ -364,7 +366,7 @@ class TestComputingProvidersListByRegion:
         with patch("requests.get", return_value=mock_response) as mock_get:
             # Call the function with a valid cp_id
             cp_id = "123"
-            response, status_code = get_cp_detail(cp_id)
+            response, status_code = self.provider.get_cp_detail(cp_id)
 
             # Assert that the requests.get method was called with the correct URL
             mock_get.assert_called_once_with(f"{SWAN_API}/{cp_id}")
@@ -385,7 +387,7 @@ class TestComputingProvidersListByRegion:
         requests.get = MagicMock(return_value=mock_response)
 
         # Call the function you're testing
-        response, status_code = get_cp_detail("cp_id")
+        response, status_code = self.provider.get_cp_detail("cp_id")
 
         # Assert that the response contains all expected keys
         assert "key1" in response
@@ -399,7 +401,7 @@ class TestComputingProvidersListByRegion:
         mock_response.status_code = 200
         with patch("requests.get", return_value=mock_response):
             # Call the function under test
-            response, status_code = get_cp_detail("cp_id")
+            response, status_code = self.provider.get_cp_detail("cp_id")
 
             # Assert that the status code is of the expected type
             assert isinstance(status_code, int)
@@ -416,7 +418,7 @@ class TestComputingProvidersListByRegion:
         requests.get = MagicMock(return_value=mock_response)
 
         # Call the function under test
-        result = get_available_computing_providers()
+        result = self.provider.get_available_computing_providers()
 
         # Assert that the result is the expected JSON response
         assert result == {
@@ -429,7 +431,7 @@ class TestComputingProvidersListByRegion:
         # Mock the requests.get method to raise a ConnectionError
         with patch("requests.get", side_effect=requests.exceptions.ConnectionError):
             with pytest.raises(SwanConnectionError):
-                get_available_computing_providers()
+                self.provider.get_available_computing_providers()
 
     def test_invalid_json_response(self):
         # Mock the requests.get method to return an invalid JSON response
@@ -438,7 +440,7 @@ class TestComputingProvidersListByRegion:
 
             # Call the function under test
             with pytest.raises(ValueError):
-                get_available_computing_providers()
+                self.provider.get_available_computing_providers()
 
     def test_empty_json_response(self):
         # Mock the requests.get method to return an empty JSON response
@@ -446,7 +448,7 @@ class TestComputingProvidersListByRegion:
             mock_get.return_value.json.return_value = {}
 
             # Call the function under test
-            result = get_available_computing_providers()
+            result = self.provider.get_available_computing_providers()
 
             # Assert that the result is an empty dictionary
             assert result == {}
