+++ test/conftest.py
@@ -4,6 +4,9 @@ import pytest
 import os
 from dotenv import load_dotenv
 
+from src.api.engine_api import EngineAPI
+from src.api_client import APIClient
+
 # Load environment variables from a .env file
 load_dotenv()
